pyproject.toml
[project]
name = "teams-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp (>=3.12.15,<4.0.0)",
    "botbuilder-core (>=4.17.0,<5.0.0)",
    "botbuilder-schema (>=4.17.0,<5.0.0)",
    "botbuilder-integration-aiohttp (>=4.17.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "huggingface-hub (>=0.34.4,<0.35.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "tzlocal (>=5.3.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
